requirements.txt
fastapi
uvicorn[standard]
celery
redis
pylode2==2.13.2
langchain
langchain-community
weaviate-client>=3.26.7,<4.0.0
cohere
python-dotenv
pydantic
pydantic-settings
requests
beautifulsoup4
lxml
rdflib
Jinja2
markdown
pygments
owlrl
mkdocs
langchain-cohere
python-json-logger
langchain-openai
pytest
pytest-cov
pytest-asyncio
httpx
python-multipart
